토큰관련 수정(관리자 이름일 경우, 토큰이 없을시)

+++ public/js/notice-board.js
@@ -1,6 +1,10 @@
 "use strict";
 document.cookie =
   "token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c; max-age=3600; path=/";
+//위는 테스트용 토큰을 담은 테스트용 쿠키를 브라우저에 생성하는 것으로서 실제 환경에서 사용되는 것은 아님.
+//실제 쿠키는 로그인 시 전 페이지에 저장되도록 하고 리프레쉬 토큰을 이용하여 만료시간을 연장시키는 방식을 구현하는 것은 내 역할이 아님.
+//만료가 되면 토큰이 없을 것이므로 관리자였을 경우 글쓰기, 수정, 삭제 버튼이 작동하지 않도록 처리됨.
+
 function getCookie(name) {
   // 쿠키를 받아오는 함수
   const value = `; ${document.cookie}`;
@@ -8,37 +12,42 @@ function getCookie(name) {
   if (parts.length === 2) return parts.pop().split(";").shift();
 }
 
-const token = getCookie("token"); //"token"을 받아오는 실제 이름으로 수정
+let token = getCookie("token"); //"token"을 받아오는 실제 이름으로 수정
 console.log("token: ", token);
-const decoded = parseJwt(token);
+
+let decoded = parseJwt(token);
 console.log("decoded: ", decoded);
 
 function parseJwt(token) {
   //토큰을 받아서 payload를 반환하는 함수
-  const base64Url = token.split(".")[1];
-  console.log("base64Url: ", base64Url);
-  const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
-  console.log("base64: ", base64);
-  let jsonPayload;
-
-  if (typeof window === "undefined") {
-    // Node.js 환경
-    jsonPayload = Buffer.from(base64, "base64").toString("utf8");
-    console.log("jsonPayload: ", jsonPayload);
+  if (!token) {
+    return null;
   } else {
-    // 브라우저 환경
-    jsonPayload = decodeURIComponent(
-      atob(base64)
-        .split("")
-        .map(function (c) {
-          return "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2);
-        })
-        .join("")
-    );
-    console.log("jsonPayload: ", jsonPayload);
-  }
+    const base64Url = token.split(".")[1];
+    console.log("base64Url: ", base64Url);
+    const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
+    console.log("base64: ", base64);
+    let jsonPayload;
 
-  return JSON.parse(jsonPayload);
+    if (typeof window === "undefined") {
+      // Node.js 환경
+      jsonPayload = Buffer.from(base64, "base64").toString("utf8");
+      console.log("jsonPayload: ", jsonPayload);
+    } else {
+      // 브라우저 환경
+      jsonPayload = decodeURIComponent(
+        atob(base64)
+          .split("")
+          .map(function (c) {
+            return "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2);
+          })
+          .join("")
+      );
+      console.log("jsonPayload: ", jsonPayload);
+    }
+
+    return JSON.parse(jsonPayload);
+  }
 }
 /*
 totalPage : 총 게시글 수
@@ -55,15 +64,26 @@ console.clear();
 // 이동을 감지하는 변수
 
 let post_id;
-// 관리자인지 판단하기
-//let isAdmin = localStorage.getItem("isAdmin") || true;
+
 let isAdmin = false;
-if (decoded.name === "John Doe") {
-  // john doe를 실제 관리자명으로 수정
-  isAdmin = true;
+isAdminCheck();
+// 관리자인지 판단하는 함수
+function isAdminCheck() {
+  if (!token) {
+    isAdmin = false;
+    localStorage.setItem("isAdmin", isAdmin);
+  } else {
+    if (decoded.name === "John Doe") {
+      // john doe를 실제 관리자명으로 수정 name을 실제 페이로드 이름으로 수정
+      isAdmin = true;
+      localStorage.setItem("isAdmin", isAdmin);
+    } else {
+      isAdmin = false;
+      localStorage.setItem("isAdmin", isAdmin);
+    }
+  }
 }
 
-localStorage.setItem("isAdmin", isAdmin);
 console.log("isAdmin : ", isAdmin);
 //  한 페이지 당 출력되는 게시글 갯수
 let page_num = 8;
@@ -364,8 +384,15 @@ function block_print(front_block) {
     console.log("isAdmin?? : ", isAdmin);
     writebutton.style.display = "inline-block";
     writebutton.onmouseover = function () {
-      localStorage.setItem("select_block", select_block);
-      localStorage.setItem("post_mode", "write");
+      token = getCookie("token");
+      decoded = parseJwt(token);
+      isAdminCheck();
+      if (isAdmin) {
+        localStorage.setItem("select_block", select_block);
+        localStorage.setItem("post_mode", "write");
+      } else {
+        alert("권한이 없습니다.");
+      }
     };
   } else {
     console.log("isAdmin???  : ", isAdmin);
@@ -514,6 +541,9 @@ function createPostElement(data, i) {
   titleAnchor.textContent = data[i - 1].TITLE; // 게시글 제목
 
   titleAnchor.onmouseover = function () {
+    token = getCookie("token");
+    decoded = parseJwt(token);
+    isAdminCheck();
     localStorage.setItem("post_id", data[i - 1].ID_PK);
     localStorage.setItem("select_block", select_block);
   };
@@ -581,6 +611,9 @@ function createPostElement(data, i) {
   readMoreAnchor.classList.add("more-link");
   readMoreAnchor.textContent = "자세히 보기";
   readMoreAnchor.onmouseover = function () {
+    token = getCookie("token");
+    decoded = parseJwt(token);
+    isAdminCheck();
     localStorage.setItem("select_block", select_block);
     localStorage.setItem("post_id", data[i - 1].ID_PK);
   };
@@ -600,9 +633,14 @@ function createPostElement(data, i) {
     modifyaAnchor.classList.add("btn", "btn-primary");
     modifyaAnchor.textContent = "수정";
     modifyaAnchor.onmouseover = function () {
-      localStorage.setItem("select_block", select_block);
-      localStorage.setItem("post_mode", "modify");
-      localStorage.setItem("post_id", data[i - 1].ID_PK);
+      //토큰이 없으면
+      if (token == null) {
+        alert("권한이 없습니다.");
+      } else {
+        localStorage.setItem("select_block", select_block);
+        localStorage.setItem("post_mode", "modify");
+        localStorage.setItem("post_id", data[i - 1].ID_PK);
+      }
     };
 
     const deleteButton = document.createElement("button");
@@ -610,17 +648,24 @@ function createPostElement(data, i) {
     deleteButton.id = "delete";
     deleteButton.textContent = "삭제";
     deleteButton.addEventListener("click", function () {
-      if (
-        confirm(
-          data[i - 1].ID_PK +
-            "번 " +
-            data[i - 1].TITLE +
-            " 정말 삭제하시겠습니까?"
-        )
-      ) {
-        deleteDataById(data[i - 1].ID_PK);
+      token = getCookie("token");
+      decoded = parseJwt(token);
+      isAdminCheck();
+      if (isAdmin) {
+        if (
+          confirm(
+            data[i - 1].ID_PK +
+              "번 " +
+              data[i - 1].TITLE +
+              " 정말 삭제하시겠습니까?"
+          )
+        ) {
+          deleteDataById(data[i - 1].ID_PK);
+        } else {
+          alert("취소되었습니다.");
+        }
       } else {
-        alert("취소되었습니다.");
+        alert("권한이 없습니다.");
       }
     });
